datapath: replace Pool bitmap/vec with HashMap and stable u64 IDs
Pool<T> previously stored items in a Vec<Option<T>> with a BitVec
tracking occupied slots. This caused iter() and get(n) to use
different indexing semantics: iter skipped empty slots while get()
indexed raw storage positions.

Replace the implementation with HashMap<u64, T> and a monotonically-
increasing u64 counter so that:

- Every inserted element gets a stable ID that is never reused after
  removal, eliminating the possibility of a stale ID resolving to a
  different element.
- Iteration is always dense — no empty slots are visited.
- get(id) and iter() are consistent: both operate on live elements.

The bit-vec crate dependency is no longer needed and is removed from
the workspace. All callers updated to use u64 IDs directly, removing
the as usize / as u64 casts that papered over the semantic mismatch.

Refs: #pool-refactor

+++ data-plane/core/datapath/src/message_processing.rs
@@ -1951,7 +1951,7 @@ mod tests {
         let rx = processor.internal.sub_ack_manager.register(sub_id);
 
         // Drop the remote connection so send_msg fails on retry.
-        processor.connection_table().remove(remote_conn as usize);
+        processor.connection_table().remove(remote_conn);
 
         let proc_clone = processor.clone();
         let handle = tokio::spawn(crate::subscription_ack::retry_loop(
